Feature#30: configure log for historic-prices-provider

+++ misc/build_lambda_zip.sh
@@ -3,21 +3,26 @@
 build_zip_archive() {
   ARCHIVE_NAME=$1
 
-  mkdir package
+  if test -f requirements.txt; then
+    echo "Creating zip archive WITH dependencies"
+    mkdir package
 
-  echo "Installing dependencies into 'package' folder"
-  python3 -m pip install --target ./package -r requirements.txt
+    echo "Installing dependencies into 'package' folder"
+    python3 -m pip install --target ./package -r requirements.txt
 
-#  chmod 755 package
+    #  chmod 755 package
 
-  cd package
-  zip -r ../$ARCHIVE_NAME.zip .
+    cd package
+    zip -r ../$ARCHIVE_NAME.zip .
 
-  cd ..
-  zip -r $ARCHIVE_NAME.zip src
+    cd ..
+    zip -r $ARCHIVE_NAME.zip src
+  else
+    echo "Creating zip archive WITHOUT dependencies"
+    zip -r $ARCHIVE_NAME.zip src
+  fi
 
   mv $ARCHIVE_NAME.zip ./../target
-
   ls -l ./../target
 }
 
@@ -28,6 +33,7 @@ build() {
   echo "Running tests"
   python3 -m unittest
 
+  echo "Building zip archive"
   build_zip_archive $1
 }
